Move routing logic to new Router class
Refactored ActivityPub routing, template, and rewrite rule logic from Activitypub class into a new Router class for better separation of concerns. Updated hooks in activitypub.php to use Router, deprecated old methods, and moved related tests to a new Test_Router class.
